Better to support large log file when -D_FILE_OFFSET_BITS=64 on 32-bit host

+++ muduo/base/FileUtil.h
@@ -74,7 +74,7 @@ class AppendFile : boost::noncopyable
 
   void flush();
 
-  size_t writtenBytes() const { return writtenBytes_; }
+  off_t writtenBytes() const { return writtenBytes_; }
 
  private:
 
@@ -82,7 +82,7 @@ class AppendFile : boost::noncopyable
 
   FILE* fp_;
   char buffer_[64*1024];
-  size_t writtenBytes_;
+  off_t writtenBytes_;
 };
 }
